requirements.txt
APScheduler==3.0.0
Flask
Flask-Compress
Flask-Markdown
Flask-Session
Flask-SocketIO==5.1.1
Jinja2
MarkupSafe
SecretStorage
Werkzeug
beautifulsoup4
discogs_client
gunicorn
hiredis
idna
imutils
itsdangerous
keyring
keyrings.alt
numpy
pandas
peewee
protobuf
psycopg2
pyOpenSSL
pycrypto
python-dateutil
python-dotenv
python-engineio
python-socketio==5.4.1
pyxdg
redis
requests>=2.26.0
simplejson
simple-websocket
six
socketIO-client
spotipy
urllib3>=1.26.7
uwsgi
websocket-client
websockets
